+++ CHANGELOG.md
@@ -19,6 +19,9 @@ available [on GitHub][2].
 - [#31](https://github.com/chshersh/zbg/pull/31):
   Command `zbg push -f` now uses git's `--force-with-lease` because it's a safer default
   (by [@kephas])
+- [#16](https://github.com/chshersh/zbg/issues/16):
+  `zbg status` now highlights staged files by making the file name bold and green.
+  (by [@leedsjohn])
 
 ## [0.2.0] — 2023-12-17 🎄